+++ apps/client/src/features/editors/Editor.tsx
@@ -6,8 +6,8 @@ import UploadModal from '../../common/components/upload-modal/UploadModal';
 import MenuBar from '../menu/MenuBar';
 import AboutModal from '../modals/about-modal/AboutModal';
 import IntegrationModal from '../modals/integration-modal/IntegrationModal';
-import ModalManager from '../modals/ModalManager';
 import QuickStart from '../modals/quick-start/QuickStart';
+import Settings from '../modals/settings/Settings';
 
 import styles from './Editor.module.scss';
 
@@ -40,7 +40,7 @@ export default function Editor() {
         <UploadModal onClose={onUploadModalClose} isOpen={isUploadModalOpen} />
         <IntegrationModal onClose={onIntegrationModalClose} isOpen={isIntegrationModalOpen} />
         <AboutModal onClose={onAboutModalClose} isOpen={isAboutModalOpen} />
-        <ModalManager isOpen={isSettingsOpen} onClose={onSettingsClose} />
+        <Settings isOpen={isSettingsOpen} onClose={onSettingsClose} />
       </ErrorBoundary>
       <div className={styles.mainContainer} data-testid='event-editor'>
         <Box id='settings' className={styles.settings}>
